aproval of pupils and showing marks for teachers and pupils are now only available if it is their course

+++ web_for_msu_front/src/pages/courses/approvePupils/ApprovePupils.jsx
@@ -16,37 +16,47 @@ const ApprovePupils = ({courseId}) => {
   useEffect(() => {
     const getAllPupilsOnCourseFunc = async () => {
       const data = await getAllPupilsOnCourse({courseId});
+      setPupils(data)
     }
 
     getAllPupilsOnCourseFunc();
   }, []);
 
+  if (!pupils || pupils.length === 0) {
+    return <></>
+  }
+
   return (
     <section style={{
       width: '100%',
+      display: 'flex',
+      flexDirection: 'column',
+      alignItems: 'center',
     }}>
       <h2 style={{
         texAlign: 'center',
+        alignSelf: 'center',
+        margin: '2rem 0',
       }}>Список учеников, ожидающих принятие на курс</h2>
 
-      {pupils && pupils.length > 0 &&
-        pupils.map((pupil) => (
-          <div key={pupil.id} style={{
-            display: 'flex',
-            justifyContent: 'space-between',
-          }}>
-            <h3>{pupil.name}</h3>
+      {pupils.map((pupil) => (
+        <div key={pupil.id} style={{
+          width: '100%',
+          display: 'flex',
+          justifyContent: 'space-between',
+        }}>
+          <h3>{pupil.name}</h3>
 
-            <div style={{
-              display: 'flex',
-              width: '40%',
-              gap: '0 2rem',
-            }}>
-              <ButtonSubmit text='Принять'/>
-              <ButtonSubmit text='Не принять' type='delete'/>
-            </div>
+          <div style={{
+            display: 'flex',
+            width: '40%',
+            gap: '0 2rem',
+          }}>
+            <ButtonSubmit text='Принять'/>
+            <ButtonSubmit text='Не принять' type='delete'/>
           </div>
-        ))}
+        </div>
+      ))}
     </section>
   );
 };
